fix: +check for existence *.new

+++ files/confpkg/doinst.sh
@@ -3,6 +3,9 @@
 config() {
     OLD=$1
     NEW=$OLD.new
+
+    [[ ! -e $NEW ]] && return 1
+
     if [[ ! -r $OLD ]]; then
         mv "$NEW" "$OLD"
     elif [[ "$(cat "$OLD" | md5sum)" = "$(cat "$NEW" | md5sum)" ]]; then
